+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.5
+
+* With this update, all OverlayLocation properties can now be used in conjunction with all AnimationDirection properties.
+* It is now possible to define a curve and animation duration for animations.
+* The barrier color can be customized.
+
 ## 0.1.1
 
 * Fixed overlay width issues on OverlayLocation.on, bottom and top positions.